Move tests to be in the same location as the EOS plugin
