+++ src/tests/helpers.test.ts
@@ -1,6 +1,6 @@
 // How to use fetch mocks - https://www.npmjs.com/package/jest-fetch-mock
-import { ZERO_HEX, EMPTY_HEX, ZERO_ADDRESS } from '../ethConstants'
-import { toWeiBN, toEthereumAddress, toEthereumPrivateKey, isNullOrEmptyEthereumValue } from '../helpers'
+import { ZERO_HEX, EMPTY_HEX, ZERO_ADDRESS } from '../plugin/ethConstants'
+import { toWeiBN, toEthereumAddress, toEthereumPrivateKey, isNullOrEmptyEthereumValue } from '../plugin/helpers'
 
 describe('Ethereum Helper Functions', () => {
   // sets fetchMock to throw an error on the next call to fetch (jsonRpc.get_abi calls fetch and triggers the error to be thrown)
